+++ example/lib/policies/gdpr_dialog.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:url_launcher/url_launcher.dart';
 
 class GdprDialog extends StatelessWidget {
-  const GdprDialog({Key? key}) : super(key: key);
+  const GdprDialog({super.key});
 
   @override
   Widget build(BuildContext context) {